continue refactor and test effort
